Split into more files for clearity
Semi-working
